backend/AI/services/document_validator/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow>=10.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
